wip: refactor to headless features
