+++ packages/app-headless-cms/src/features/graphQLClient/abstractions.ts
